grt: use int instead of long

+++ src/grt/include/grt/GlobalRouter.h
@@ -158,7 +158,7 @@ class GlobalRouter
                            float reduction_percentage);
   void setVerbose(const bool v);
   void setOverflowIterations(int iterations);
-  void setGridOrigin(long x, long y);
+  void setGridOrigin(int x, int y);
   void setAllowCongestion(bool allow_congestion);
   void setMacroExtension(int macro_extension);
   void printGrid();
